requirements.txt
requests==2.28.1
msal==1.18.0
tqdm==4.65.0
python-dotenv==0.21.0
